extract duplicate test helpers into shared scope
