+++ FeedFeatureTests/Feed API/URLSessionHTTPClientTests.swift
@@ -122,10 +122,6 @@ class URLSessionHTTPClientTests: XCTestCase {
         return sut
     }
 
-    private func anyURL() -> URL {
-        return URL(string: "https://example.com")!
-    }
-
     private func anyData() -> Data {
         return Data("any data".utf8)
     }
@@ -136,10 +132,6 @@ class URLSessionHTTPClientTests: XCTestCase {
 
     private func anyHTTPURLResponse() -> HTTPURLResponse {
         return HTTPURLResponse(url: anyURL(), statusCode: 200, httpVersion: nil, headerFields: nil)!
-    }
-
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any error", code: 0)
     }
 
     private class URLProtocolStub: URLProtocol {
