+++ FeedFeatureTests/Feed Cache/ValidateFeedCacheUseCaseTests.swift
@@ -41,10 +41,6 @@ class ValidateFeedCacheUseCaseTests: XCTestCase {
 
     // MARK: Helpers
 
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any error", code: 0)
-    }
-
     private func makeSUT(currentDate: @escaping () -> Date = Date.init ,file: StaticString = #filePath, line: UInt = #line) -> (sut: LocalFeedLoader, store: FeedStoreSpy) {
         let store = FeedStoreSpy()
         let sut = LocalFeedLoader(store: store, currentDate: currentDate)
@@ -52,30 +48,6 @@ class ValidateFeedCacheUseCaseTests: XCTestCase {
         trackForMemoryLeaks(store,file: file, line: line)
         return (sut, store)
     }
-
-    private func uniqueItems() -> (models: [FeedItem], local: [LocalFeedItem]) {
-        let items = [uniqueItem(), uniqueItem()]
-        let local = items.map { LocalFeedItem(id: $0.id, description: $0.description, location: $0.location, imageURL: $0.imageURL) }
-        return (items, local)
-    }
-
-    private func uniqueItem() -> FeedItem {
-        return FeedItem(id: UUID(), description: "any", location: "any", imageURL: anyURL())
-    }
-
-    private func anyURL() -> URL {
-        return URL(string: "https://example.com")!
-    }
-
-
 }
 
-private extension Date {
-    func adding(days: Int) -> Date {
-        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-    }
 
-    func adding(seconds: TimeInterval) -> Date {
-        return self + seconds
-    }
-}
